FEAT: seeder migration
1. roles
2. permissions
3. role_permissions
4. users
