+++ app/models/multi_tenancy.py
@@ -1,5 +1,5 @@
-from sqlalchemy import Column, Integer, String, DateTime, Float, ForeignKey
-import datetime
+from sqlalchemy import Column, Integer, String, TIMESTAMP, Float, ForeignKey
+from sqlalchemy.sql import func
 from app.models.base import Base, TenantMixin
 
 class Note(Base, TenantMixin):
@@ -9,7 +9,11 @@ class Note(Base, TenantMixin):
     id = Column(Integer, primary_key=True, index=True)
     title = Column(String)
     description = Column(String)
-    created_at = Column(DateTime, default=datetime.datetime.utcnow)
+
+    created_at = Column(TIMESTAMP(timezone=True), server_default=func.now())
+    created_by = Column(String(255), nullable=True)
+    updated_at = Column(TIMESTAMP(timezone=True), server_default=func.now())
+    updated_by = Column(String(255), nullable=True)
 
 class Order(Base, TenantMixin):
     __tablename__ = 'orders'
@@ -21,4 +25,8 @@ class Order(Base, TenantMixin):
     note_id = Column(Integer, ForeignKey("tenant_schema.notes.id"), nullable=True)
     quantity = Column(Integer, nullable=False)
     total_price = Column(Float, nullable=False)
-    created_at = Column(DateTime, default=datetime.datetime.utcnow)
+
+    created_at = Column(TIMESTAMP(timezone=True), server_default=func.now())
+    created_by = Column(String(255), nullable=True)
+    updated_at = Column(TIMESTAMP(timezone=True), server_default=func.now())
+    updated_by = Column(String(255), nullable=True)
