+++ app/models/permissions.py
@@ -1,4 +1,5 @@
-from sqlalchemy import Column, Integer, String
+from sqlalchemy import Column, Integer, String, TIMESTAMP
+from sqlalchemy.sql import func
 from app.core.database import Base
 
 class Permission(Base):
@@ -7,3 +8,8 @@ class Permission(Base):
 
     id = Column(Integer, primary_key=True, index=True)
     name = Column(String(255), unique=True, nullable=False)
+
+    created_at = Column(TIMESTAMP(timezone=True), server_default=func.now())
+    created_by = Column(String(255), nullable=True)
+    updated_at = Column(TIMESTAMP(timezone=True), server_default=func.now())
+    updated_by = Column(String(255), nullable=True)
