--- app/models/product.py
@@ -1,13 +1,0 @@
-from sqlalchemy import Column, Integer, String, Float, DateTime
-import datetime
-from app.core.database import Base
-
-class Product(Base):
-    __tablename__ = "products"
-    __table_args__ = {"schema": "public"}
-
-    id = Column(Integer, primary_key=True, index=True)
-    name = Column(String, index=True)
-    price = Column(Float, nullable=False)
-    stock = Column(Integer, nullable=False)
-    created_at = Column(DateTime, default=datetime.datetime.utcnow)
