+++ app/models/products.py
@@ -1,13 +1,15 @@
-from sqlalchemy import Column, Integer, String, TIMESTAMP
+from sqlalchemy import Column, Integer, String, Float, TIMESTAMP
 from sqlalchemy.sql import func
 from app.core.database import Base
 
-class Role(Base):
-    __tablename__ = "roles"
+class Product(Base):
+    __tablename__ = "products"
     __table_args__ = {"schema": "public"}
 
     id = Column(Integer, primary_key=True, index=True)
-    name = Column(String, unique=True, index=True)
+    name = Column(String, index=True)
+    price = Column(Float, nullable=False)
+    stock = Column(Integer, nullable=False)
 
     created_at = Column(TIMESTAMP(timezone=True), server_default=func.now())
     created_by = Column(String(255), nullable=True)
