+++ app/models/role_permissions.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, ForeignKey, TIMESTAMP, String
+from sqlalchemy import Column, Integer, TIMESTAMP, String
 from sqlalchemy.sql import func
 from app.core.database import Base
 
@@ -12,5 +12,5 @@ class RolePermission(Base):
 
     created_at = Column(TIMESTAMP(timezone=True), server_default=func.now())
     created_by = Column(String(255), nullable=True)
-    updated_at = Column(TIMESTAMP(timezone=True), nullable=True)
+    updated_at = Column(TIMESTAMP(timezone=True), server_default=func.now())
     updated_by = Column(String(255), nullable=True)
